refactoring: file splited and worked file trnsfer with two pync. other use casees are not tested yet.
